Refactored several file names and changed file name formatting and training set lines.
